requirements.txt
pytest
allure-pytest
selene>=2.0.0rc4
Appium-Python-Client==4.3.0
